(長田)シフトカレンダーの閉じるボタン及びデザイン
